Changed InsertResponse to return an Integer, which is the position that the exercise is inserted to. Adapter is notified of this position and updates every view after or on that position.

+++ app/src/main/java/com/longlife/workoutlogger/v2/view/ExercisesOverview/ExerciseCreateFragment.java
@@ -114,7 +114,7 @@ public class ExerciseCreateFragment extends Fragment {
     ///
     /// INSERT EXERCISE RENDERING
     ///
-    private void processInsertResponse(Response<Long> response) {
+    private void processInsertResponse(Response<Integer> response) {
         switch (response.getStatus()) {
             case LOADING:
                 renderLoadingState();
@@ -128,7 +128,7 @@ public class ExerciseCreateFragment extends Fragment {
         }
     }
 
-    private void renderSuccessState(Long id) {
+    private void renderSuccessState(Integer id) {
         StringBuilder sb = new StringBuilder();
         sb.append("inserted exercise ");
         sb.append(id.toString());
